Don't truncate explicit entry coefficients to int
MatrixEntry::coeff and the row-major form are long, but setEntries passed them
through static_cast<int>, silently truncating anything outside int range.
ARingZZp and ARingQQFlint both provide set(ElementType&, long), so pass the
long straight through.

Also spell the integer one as 1L, selecting that same overload directly rather
than via the int one.

+++ M2/Macaulay2/e/unit-tests/ARingMatrixTest.hpp
@@ -159,7 +159,7 @@ class ARingMatrixGenerator
     for (const auto& e : entries)
       {
         assert(e.row < M.numRows() && e.col < M.numColumns());
-        mRing.set(a, static_cast<int>(e.coeff));
+        mRing.set(a, e.coeff);
         Setter::set(M, e.row, e.col, a);
       }
   }
@@ -173,7 +173,7 @@ class ARingMatrixGenerator
     size_t i = 0;
     for (long v : values)
       {
-        mRing.set(a, static_cast<int>(v));
+        mRing.set(a, v);
         Setter::set(M, i / M.numColumns(), i % M.numColumns(), a);
         i++;
       }
@@ -229,14 +229,14 @@ class ARingMatrixGenerator
         mElements.nextElement(a);
         if (!mRing.is_zero(a)) break;
       }
-    if (mRing.is_zero(a)) mRing.set(a, 1);
+    if (mRing.is_zero(a)) mRing.set(a, 1L);
     Setter::set(M, r, c, a);
   }
 
   void fillIdentity(MatType& M)
   {
     Element one(mRing);
-    mRing.set(one, 1);
+    mRing.set(one, 1L);
     for (size_t i = 0; i < M.numRows(); i++) Setter::set(M, i, i, one);
   }
 
@@ -290,13 +290,13 @@ class ARingMatrixGenerator
         for (size_t i = 0; i < nrows; i++)
           {
             if (i < k) mRing.set_zero(u[k * nrows + i]);
-            else if (i == k) mRing.set(u[k * nrows + i], 1);
+            else if (i == k) mRing.set(u[k * nrows + i], 1L);
             else mElements.nextElement(u[k * nrows + i]);
           }
         for (size_t j = 0; j < ncols; j++)
           {
             if (j < k) mRing.set_zero(v[k * ncols + j]);
-            else if (j == k) mRing.set(v[k * ncols + j], 1);
+            else if (j == k) mRing.set(v[k * ncols + j], 1L);
             else mElements.nextElement(v[k * ncols + j]);
           }
       }
